feat:add busStop in maps

+++ app/src/main/java/com/gilbersoncampos/testeaiko/data/remote/dto/BusStopResponseDto.kt
@@ -7,6 +7,6 @@ data class BusStopResponseDto(
     val cp:Long,
     val np:String,
     val ed: String,
-    val py: Float,
-    val px: Float,
+    val py: Double,
+    val px: Double,
 )
